sessionerne virker! Brug UserViewDTO til ALLE sessions i controller, og giv dem som parameter i service-klassens updateUser() og createUser()
