+++ src/main/java/com/example/estimationtool/controller/UserController.java
@@ -1,10 +1,11 @@
-package com.example.estimationtool.user;
+package com.example.estimationtool.controller;
 
-import com.example.estimationtool.dto.LoginDTO;
 import com.example.estimationtool.dto.UserRegistrationDTO;
 import com.example.estimationtool.dto.UserUpdateDTO;
 import com.example.estimationtool.dto.UserViewDTO;
 import com.example.estimationtool.enums.Role;
+import com.example.estimationtool.model.User;
+import com.example.estimationtool.service.UserService;
 import jakarta.servlet.http.HttpSession;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -24,39 +25,21 @@ public class UserController {
         this.userService = userService;
     }
 
-    @GetMapping("/login")
-    public String getLogin() {
-        return "index";
+
+    private UserViewDTO getCurrentUser(HttpSession session) {
+        return (UserViewDTO) session.getAttribute("currentUser");
     }
 
-    @GetMapping("/logout")
-    public String logout(HttpSession session) {
-        session.invalidate();
-        return "index";
-    }
-
-    @PostMapping("/login")
-    public String postLogin(@RequestParam("email") String email,
-                            @RequestParam("password") String password,
-                            HttpSession session,
-                            RedirectAttributes redirectAttributes) {
-
-        UserViewDTO userViewDTO = userService.login(email, password); // Kaster exception ved fejl fra @Service
-
-
-        session.setAttribute("user", userViewDTO);
-        session.setMaxInactiveInterval(1000);
-        redirectAttributes.addFlashAttribute("success", "Du er nu logget ind.");
-
-        return "redirect:/users/profile";
-    }
 
     @GetMapping("/profile")
     public String getFrontPage(HttpSession session,
                                RedirectAttributes redirectAttributes,
                                Model model) {
 
-        UserViewDTO currentUser = (UserViewDTO) session.getAttribute("user");
+        UserViewDTO currentUser = getCurrentUser(session);
+
+
+        //UserViewDTO currentUser = (UserViewDTO) session.getAttribute("currentUser");
 
         if (currentUser == null) {
             redirectAttributes.addFlashAttribute("error", "Du skal være logget ind for at tilgå forsiden.");
@@ -64,7 +47,7 @@ public class UserController {
         }
 
         model.addAttribute("user", currentUser); // valgfrit – Thymeleaf kan også hente fra session direkte
-        return "front-page";
+        return "user/front-page";
     }
 
 
@@ -75,7 +58,9 @@ public class UserController {
                                  Model model,
                                  RedirectAttributes redirectAttributes) {
 
-        User currentUser = (User) session.getAttribute("currentUser");
+        UserViewDTO currentUser = getCurrentUser(session);
+
+       // User currentUser = (User) session.getAttribute("currentUser");
 
         if (currentUser == null) {
             redirectAttributes.addFlashAttribute("error", "Log ind for at oprette en ny bruger.");
@@ -94,7 +79,10 @@ public class UserController {
                              HttpSession session,
                              RedirectAttributes redirectAttributes) {
 
-        User currentUser = (User) session.getAttribute("currentUser");
+        UserViewDTO currentUser = getCurrentUser(session);
+
+
+        //User currentUser = (User) session.getAttribute("currentUser");
 
         // Tjekker om brugeren er logget ind
         if (currentUser == null) {
@@ -114,12 +102,15 @@ public class UserController {
 
     //------------------------------------ Read() --------------------------------------
 
-    @GetMapping("")
+    @GetMapping("users")
     public String showAllUsers(HttpSession session,
                                Model model,
                                RedirectAttributes redirectAttributes) {
 
-        User currentUser = (User) session.getAttribute("currentUser");
+        UserViewDTO currentUser = getCurrentUser(session);
+
+        //User currentUser = (User) session.getAttribute("currentUser");
+
 
         // Tjekker om brugeren er logget ind
         if (currentUser == null) {
@@ -137,7 +128,10 @@ public class UserController {
                            Model model,
                            RedirectAttributes redirectAttributes) {
 
-        User currentUser = (User) session.getAttribute("currentUser");
+        UserViewDTO currentUser = getCurrentUser(session);
+
+
+        //User currentUser = (User) session.getAttribute("currentUser");
 
         // Tjekker om brugeren er logget ind
         if (currentUser == null) {
@@ -159,7 +153,10 @@ public class UserController {
                                Model model,
                                RedirectAttributes redirectAttributes) {
 
-        User currentUser = (User) session.getAttribute("currentUser");
+        UserViewDTO currentUser = getCurrentUser(session);
+
+
+        //User currentUser = (User) session.getAttribute("currentUser");
 
         // Tjekker om bruger er logget ind
         if (currentUser == null) {
@@ -193,7 +190,10 @@ public class UserController {
                              HttpSession session,
                              RedirectAttributes redirectAttributes) {
 
-        User currentUser = (User) session.getAttribute("currentUser");
+        UserViewDTO currentUser = getCurrentUser(session);
+
+
+        //User currentUser = (User) session.getAttribute("currentUser");
 
         // Tjekker om bruger er logget ind
         if (currentUser == null) {
