+++ src/main/java/com/example/estimationtool/repository/UserRepository.java
@@ -1,8 +1,8 @@
-package com.example.estimationtool.user;
+package com.example.estimationtool.repository;
 
-import com.example.estimationtool.dto.UserViewDTO;
 import com.example.estimationtool.interfaces.IUserRepository;
-import org.springframework.dao.DataAccessException;
+import com.example.estimationtool.model.User;
+import com.example.estimationtool.rowMapper.UserRowMapper;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
