--- src/main/resources/templates/front-page.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-</head>
-<body>
-<h1>Det virker</h1>
-</body>
-</html>
