+++ src/main/resources/templates/index.html
@@ -16,7 +16,7 @@
 </div>
 
 
-<form th:action="@{/users/login}" method="post">
+<form th:action="@{/login}" method="post">
     <div>
         <label for="email">Email:</label>
         <input type="email" id="email" name="email" required/>
