+++ src/test/java/com/example/estimationtool/unitTest/controller/UserControllerTest.java
@@ -4,9 +4,9 @@ package com.example.estimationtool.unitTest.controller;
 import com.example.estimationtool.dto.UserRegistrationDTO;
 import com.example.estimationtool.dto.UserViewDTO;
 import com.example.estimationtool.enums.Role;
-import com.example.estimationtool.user.User;
-import com.example.estimationtool.user.UserController;
-import com.example.estimationtool.user.UserService;
+import com.example.estimationtool.model.User;
+import com.example.estimationtool.controller.UserController;
+import com.example.estimationtool.service.UserService;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,11 +15,8 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.web.servlet.MockMvc;
 
-import org.springframework.ui.Model;
-
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
