+++ src/test/java/com/example/estimationtool/unitTest/service/UserServiceTest.java
@@ -3,9 +3,8 @@ package com.example.estimationtool.unitTest.service;
 import com.example.estimationtool.enums.Role;
 import com.example.estimationtool.dto.UserRegistrationDTO;
 import com.example.estimationtool.interfaces.IUserRepository;
-import com.example.estimationtool.user.User;
-import com.example.estimationtool.user.UserService;
-import org.junit.jupiter.api.BeforeEach;
+import com.example.estimationtool.model.User;
+import com.example.estimationtool.service.UserService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
